requirements.txt
Flask==2.0.1
pyTelegramBotAPI==4.12.0
flask-cors==4.0.0 